Xinder :: Re-Worked.
First off, wake up! It's the first of the month. Anyway! I have re-worked Xinder to try and make sure that it is a full (three-part project) following the likes of xFlix & xGram (only two-parts). To feel like a full project and feel a little more in-depth and more complex than what it was before.